requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-wtf
asgiref
blinker
click
colorama
itsdangerous
Jinja2
MarkupSafe
mysql-connector-python
Werkzeug
SpeechRecognition
xhtml2pdf